classes don't jump semesters any more

+++ CoursePlanner/Controllers/CoursePlanController.cs
@@ -32,32 +32,36 @@ namespace CoursePlanner.Controllers
                 ClientCourseInfo info = JsonConvert.DeserializeObject<ClientCourseInfo>(p.ToString());
                 planList.plan.Add(info);
             });
-
-            int studentId = planList.plan[0].studentId;
-
-            if (!ModelState.IsValid)
+            // only save plan if there is content within it
+            if (planList.plan.Count > 0)
             {
-                return BadRequest(ModelState);
-            }
-            string connectionString = "Data Source=(LocalDb)\\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\\aspnet-CoursePlanner-20180131110323.mdf;Initial Catalog=aspnet-CoursePlanner-20180131110323;Integrated Security=True";
-            int success;
-            using (SqlConnection conn = new SqlConnection(connectionString))
-            {
-                conn.Open();
-                // remove old classes
-                SqlCommand deleteCommand = new SqlCommand("DELETE from student_course WHERE student_id=@0", conn);
-                deleteCommand.Parameters.Add(new SqlParameter("0", studentId));
-                deleteCommand.ExecuteNonQuery();
+                int studentId = planList.plan[0].studentId;
 
-                // add new configuration
-                planList.plan.ForEach(p => {
-                    SqlCommand command = new SqlCommand("INSERT into student_course (course_id, student_id, semester_id) values (@0, @1, @2)", conn);
-                    command.Parameters.Add(new SqlParameter("0", p.courseId));
-                    command.Parameters.Add(new SqlParameter("1", p.studentId));
-                    command.Parameters.Add(new SqlParameter("2", p.semesterId));
-                    success = command.ExecuteNonQuery();
-                });
-                conn.Close();
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
+                string connectionString = "Data Source=(LocalDb)\\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\\aspnet-CoursePlanner-20180131110323.mdf;Initial Catalog=aspnet-CoursePlanner-20180131110323;Integrated Security=True";
+                int success;
+                using (SqlConnection conn = new SqlConnection(connectionString))
+                {
+                    conn.Open();
+                    // remove old classes
+                    SqlCommand deleteCommand = new SqlCommand("DELETE from student_course WHERE student_id=@0", conn);
+                    deleteCommand.Parameters.Add(new SqlParameter("0", studentId));
+                    deleteCommand.ExecuteNonQuery();
+
+                    // add new configuration
+                    planList.plan.ForEach(p =>
+                    {
+                        SqlCommand command = new SqlCommand("INSERT into student_course (course_id, student_id, semester_id) values (@0, @1, @2)", conn);
+                        command.Parameters.Add(new SqlParameter("0", p.courseId));
+                        command.Parameters.Add(new SqlParameter("1", p.studentId));
+                        command.Parameters.Add(new SqlParameter("2", p.semesterId));
+                        success = command.ExecuteNonQuery();
+                    });
+                    conn.Close();
+                }
             }
             return Ok();
         }
